FT-109: Fix crash after editing amount on AddAccountScreen
