+++ common/src/commonMain/kotlin/com/finance_tracker/finance_tracker/presentation/add_transaction/views/AddButtonSection.kt
@@ -72,7 +72,7 @@ fun ActionButtonsSection(
                         onAddClick.invoke()
                     }
                 },
-                enabled = enabled
+                enable = enabled
             )
         }
     }
